handle zipped file with name differing from folder
`heudiconv` looks for the DICOMs in ${LOCAL_DICOM}, and ${LOCAL_DICOM} is removed after conversion. But this previously assumed that LOCAL_DICOM matched the name of a folder that was inside ${FILES}. That match doesn't always happen (e.g., in the zipped file for a traveling human). This patch unzips to a new folder called LOCAL_DICOM, so LOCAL_DICOM will always exist for `heudiconv` and `rm`